Don't change the Z-RAM Date
a2-4am/4cade#346 Currently the FASTChip is not available for the //c and its addresses are the same as in the Z-RAM Ultra 2/3 Clock. The FASTChip activation/deactivation modifies the date and sets it to invalid values that prevent ProDOS 2.4.2 from booting.
To avoid this, if a //c is detected a shorter addon routine only with ZipChip support is used.