skip playing Pollywog demo on ROM0 or ROM1 IIgs
Better fix; just don't play the Pollywog demo on a ROM0 or ROM1 that can't properly display page 2. We can't turn on Alternate Display Mode due to its use of VBL interrupts on these machines. We can detect and turn it on with ROM3, however, where it does not use interrupts.